project/requirements.txt
pyspark
mysql-connector-python
pandas
pyyaml
